add Sever interface
