+++ hw12_13_14_15_calendar/internal/server/server.go
@@ -1,0 +1,23 @@
+package server
+
+import (
+	"context"
+	"time"
+
+	"github.com/google/uuid"
+	"github.com/grahovsky/go-hw/hw12_13_14_15_calendar/internal/models"
+)
+
+type Application interface {
+	AddEvent(context.Context, *models.Event) error
+	GetEvent(context.Context, uuid.UUID) (*models.Event, error)
+	GetEventsForPeriod(ctx context.Context, from, to time.Time) ([]models.Event, error)
+	ListEvents(ctx context.Context, limit, low uint64) ([]models.Event, error)
+	UpdateEvent(ctx context.Context, event *models.Event) error
+	DeleteEvent(ctx context.Context, id uuid.UUID) error
+}
+
+type Server interface {
+	Start(context.Context) error
+	Stop(context.Context) error
+}
